+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=68", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "bundestag-mine-refactor"
+name = "mine"
 version = "0.1.0"
 description = "Modernisierte Python-Pipeline für Bundestagsprotokolle"
 readme = "README.md"
@@ -21,7 +21,7 @@ dev = [
 ]
 
 [project.scripts]
-bundestag-mine-refactor = "bundestag_mine_refactor.cli:main"
+mine = "mine.cli:main"
 
 [tool.setuptools.packages.find]
 where = ["src"]
