+++ requirements.txt
@@ -1,4 +1,4 @@
-# Laufzeitabhängigkeiten für bundestag-mine-refactor
+# Laufzeitabhängigkeiten für Mine
 httpx>=0.27
 SQLAlchemy>=2.0
 google-genai>=1.38
