+++ src/mine/ui/__init__.py
@@ -1,11 +1,11 @@
-"""Interactive control center for the Bundestag Mine pipeline."""
+"""Interactive control center for the Mine pipeline."""
 from __future__ import annotations
 
 from importlib import import_module
 from typing import Any
 
 try:
-    run_ui = import_module("bundestag_mine_refactor.ui.app").run_ui  # type: ignore[attr-defined]
+    run_ui = import_module("mine.ui.app").run_ui  # type: ignore[attr-defined]
 except ModuleNotFoundError as exc:  # pragma: no cover - triggered when nicegui is absent
     if exc.name != "nicegui":
         raise
@@ -13,7 +13,7 @@ except ModuleNotFoundError as exc:  # pragma: no cover - triggered when nicegui 
     def run_ui(*_: Any, **__: Any) -> None:
         raise ModuleNotFoundError(
             "NiceGUI muss installiert sein, um die grafische Oberfläche zu starten. "
-            "Installieren Sie das Extra mit `pip install bundestag-mine-refactor[nicegui]` "
+            "Installieren Sie Mine mit `pip install mine` "
             "oder direkt `pip install nicegui>=1.4.17`."
         ) from exc
