+++ tests/clients/test_dip_client.py
@@ -2,7 +2,7 @@ from datetime import date
 
 import pytest
 
-from bundestag_mine_refactor.clients import DIPClient
+from mine.clients import DIPClient
 
 
 class DummyClient(DIPClient):
