+++ tests/parsing/test_speeches.py
@@ -1,4 +1,4 @@
-from bundestag_mine_refactor.parsing import parse_speeches
+from mine.parsing import parse_speeches
 
 
 def test_parse_speeches_extracts_multiple_segments():
